+++ src/test/kotlin/org/kalasim/examples/api/TickTrafoExample.kts
@@ -1,18 +1,32 @@
 //TickTrafoExample.kts
 import org.kalasim.*
-import java.time.Duration
 import java.time.Instant
 import java.util.concurrent.TimeUnit
 
 
 createSimulation(true) {
-    tickTransform = OffsetTransform(Instant.now(), TimeUnit.MINUTES)
+//    tickTransform = OffsetTransform(Instant.now(), TimeUnit.MINUTES)
+    tickTransform = TickTransform(TimeUnit.MINUTES)
+
+    object :Component(){
+        override fun process() =sequence<Component> {
+            hold(1.minutes, description = "dressing")
+
+            // we can express fractional durations (1.3 hours = 78 minutes)
+            hold(1.3.hours, description = "walking")
+
+            // we can also hold until a specific time
+            hold(until= now + 3.hours )
+        }
+    }
 
     // run until a specific time
-    run(until = Instant.now().plus(Duration.ofHours(1)).asTickTime())
+    run(until = Instant.now() + 1.hours)
 
     // run for a given duration
-    run(Duration.ofHours(1).asTicks())
+    run(1.days)
 
+    println(now)
     println(asWallTime(now))
 }
+